MP - done

+++ src/main/java/com/humber/model/Product.java
@@ -8,7 +8,7 @@ public class Product {
 
     @Id
     @GeneratedValue(strategy= GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     @Column(name = "name")
     private String name;
     @Column(name = "price")
@@ -21,7 +21,7 @@ public class Product {
     public Product() {
         super();
     }
-    public Product(Integer id, String name, double price, int quantity, String category) {
+    public Product(Long id, String name, double price, int quantity, String category) {
         super();
         this.id = id;
         this.name = name;
@@ -30,11 +30,11 @@ public class Product {
         this.category = category;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
